Enum rustc_ast::ast::MacArgs

pub enum MacArgs {
    Empty,
    Delimited(DelimSpan, MacDelimiter, TokenStream),
    Eq(Span, MacArgsEq),
}

Arguments passed to an attribute or a function-like macro.

Variants

Empty

No arguments - #[attr].

Delimited(DelimSpan, MacDelimiter, TokenStream)

Delimited arguments - #[attr()/[]/{}] or mac!()/[]/{}.

Eq(Span, MacArgsEq)

Tuple Fields

0: Span

Span of the = token.

1: MacArgsEq

The “value”.

Arguments of a key-value attribute - #[attr = "value"].

Implementations

impl MacArgs

pub fn delim(&self) -> Option<Delimiter>

pub fn span(&self) -> Option<Span>

pub fn inner_tokens(&self) -> TokenStream

Tokens inside the delimiters or after =. Proc macros see these tokens, for example.

pub fn need_semicolon(&self) -> bool

Whether a macro with these arguments needs a semicolon when used as a standalone item or statement.