Enum rustc_ast::ast::NestedMetaItem

pub enum NestedMetaItem {
    MetaItem(MetaItem),
    Lit(MetaItemLit),
}

Values inside meta item lists.

E.g., each of Clone, Copy in #[derive(Clone, Copy)].

Variants

MetaItem(MetaItem)

A full MetaItem, for recursive meta items.

Lit(MetaItemLit)

A literal.

E.g., "foo", 64, true.

Implementations

impl NestedMetaItem

pub fn span(&self) -> Span

pub fn ident(&self) -> Option<Ident>

For a single-segment meta item, returns its name; otherwise, returns None.

pub fn name_or_empty(&self) -> Symbol

pub fn has_name(&self, name: Symbol) -> bool

Returns true if this list item is a MetaItem with a name of name.

pub fn is_word(&self) -> bool

Returns true if self is a MetaItem and the meta item is a word.

pub fn meta_item_list(&self) -> Option<&[NestedMetaItem]>

Gets a list of inner meta items from a list MetaItem type.

pub fn name_value_literal(&self) -> Option<(Symbol, &MetaItemLit)>

Returns a name and single literal value tuple of the MetaItem.

pub fn name_value_literal_span(&self) -> Option<Span>

See MetaItem::name_value_literal_span.

pub fn value_str(&self) -> Option<Symbol>

Gets the string value if self is a MetaItem and the MetaItem is a MetaItemKind::NameValue variant containing a string, otherwise None.

pub fn lit(&self) -> Option<&MetaItemLit>

Returns the MetaItemLit if self is a NestedMetaItem::Literals.

pub fn meta_item(&self) -> Option<&MetaItem>

Returns the MetaItem if self is a NestedMetaItem::MetaItem.

pub fn is_meta_item(&self) -> bool

Returns true if the variant is MetaItem.

fn from_tokens<'a, I>(tokens: &mut Peekable<I>) -> Option<NestedMetaItem>where I: Iterator<Item = &'a TokenTree>,