Enum rustc_ast::ast::PatKind

pub enum PatKind {

    Wild,
    Ident(BindingAnnotation, Ident, Option<P<Pat>>),
    Struct(Option<QSelf>, Path, Vec<PatField>, bool),
    TupleStruct(Option<QSelf>, Path, Vec<P<Pat>>),
    Or(Vec<P<Pat>>),
    Path(Option<QSelf>, Path),
    Tuple(Vec<P<Pat>>),
    Box(P<Pat>),
    Ref(P<Pat>, Mutability),
    Lit(P<Expr>),
    Range(Option<P<Expr>>, Option<P<Expr>>, Spanned<RangeEnd>),
    Slice(Vec<P<Pat>>),
    Rest,
    Paren(P<Pat>),
    MacCall(P<MacCall>),

}

All the different flavors of pattern that Rust recognizes.

Variants

Wild

Represents a wildcard pattern (_).

Ident(BindingAnnotation, Ident, Option<P<Pat>>)

A PatKind::Ident may either be a new bound variable (ref mut binding @ OPT_SUBPATTERN), or a unit struct/variant pattern, or a const pattern (in the last two cases the third field must be None). Disambiguation cannot be done with parser alone, so it happens during name resolution.

Struct(Option<QSelf>, Path, Vec<PatField>, bool)

A struct or struct variant pattern (e.g., Variant {x, y, ..}). The bool is true in the presence of a ...

TupleStruct(Option<QSelf>, Path, Vec<P<Pat>>)

A tuple struct/variant pattern (Variant(x, y, .., z)).

Or(Vec<P<Pat>>)

An or-pattern A | B | C. Invariant: pats.len() >= 2.

Path(Option<QSelf>, Path)

A possibly qualified path pattern. Unqualified path patterns A::B::C can legally refer to variants, structs, constants or associated constants. Qualified path patterns <A>::B::C/<A as Trait>::B::C can only legally refer to associated constants.

Tuple(Vec<P<Pat>>)

A tuple pattern ((a, b)).

Box(P<Pat>)

A box pattern.

Ref(P<Pat>, Mutability)

A reference pattern (e.g., &mut (a, b)).

Lit(P<Expr>)

A literal.

Range(Option<P<Expr>>, Option<P<Expr>>, Spanned<RangeEnd>)

A range pattern (e.g., 1...2, 1..2, 1.., ..2, 1..=2, ..=2).

Slice(Vec<P<Pat>>)

A slice pattern [a, b, c].

Rest

A rest pattern ...

Syntactically it is valid anywhere.

Semantically however, it only has meaning immediately inside:

a slice pattern: [a, .., b],

a binding pattern immediately inside a slice pattern: [a, r @ ..],

a tuple pattern: (a, .., b),

a tuple struct/variant pattern: $path(a, .., b).

In all of these cases, an additional restriction applies, only one rest pattern may occur in the pattern sequences.

Paren(P<Pat>)

Parentheses in patterns used for grouping (i.e., (PAT)).

MacCall(P<MacCall>)

A macro pattern; pre-expansion.