Struct rustc_ast::ast::Attribute

pub struct Attribute {
    pub kind: AttrKind,
    pub id: AttrId,
    pub style: AttrStyle,
    pub span: Span,
}

Metadata associated with an item.

Fields
kind: AttrKindid: AttrIdstyle: AttrStyle

Denotes if the attribute decorates the following construct (outer) or the construct this attribute is contained within (inner).

span: Span
Implementations

impl Attribute

pub fn has_name(&self, name: Symbol) -> bool

pub fn ident(&self) -> Option<Ident>

For a single-segment attribute, returns its name; otherwise, returns None.

pub fn name_or_empty(&self) -> Symbol

pub fn value_str(&self) -> Option<Symbol>

pub fn meta_item_list(&self) -> Option<Vec<NestedMetaItem>>

pub fn is_word(&self) -> bool

impl Attribute

pub fn is_doc_comment(&self) -> bool

Returns true if it is a sugared doc comment (/// or //! for example). So #[doc = "doc"] (which is a doc comment) and #[doc(...)] (which is not a doc comment) will return false.

pub fn doc_str_and_comment_kind(&self) -> Option<(Symbol, CommentKind)>

Returns the documentation and its kind if this is a doc comment or a sugared doc comment.

///doc returns Some(("doc", CommentKind::Line)).

/** doc */ returns Some(("doc", CommentKind::Block)).

#[doc = "doc"] returns Some(("doc", CommentKind::Line)).

#[doc(...)] returns None.

pub fn doc_str(&self) -> Option<Symbol>

Returns the documentation if this is a doc comment or a sugared doc comment.

///doc returns Some("doc").

#[doc = "doc"] returns Some("doc").

#[doc(...)] returns None.

pub fn may_have_doc_links(&self) -> bool

pub fn get_normal_item(&self) -> &AttrItem

pub fn unwrap_normal_item(self) -> AttrItem

pub fn meta(&self) -> Option<MetaItem>

Extracts the MetaItem from inside this Attribute.

pub fn meta_kind(&self) -> Option<MetaItemKind>

pub fn tokens(&self) -> TokenStream