Struct rustc_ast::ast::Expr

pub struct Expr {
    pub id: NodeId,
    pub kind: ExprKind,
    pub span: Span,
    pub attrs: AttrVec,
    pub tokens: Option<LazyAttrTokenStream>,
}

An expression.

Fields
id: NodeIdkind: ExprKindspan: Spanattrs: AttrVectokens: Option<LazyAttrTokenStream>
Implementations

impl Expr

pub fn returns(&self) -> bool

Returns true if this expression would be valid somewhere that expects a value; for example, an if condition.

pub fn is_potential_trivial_const_param(&self) -> bool

Is this expr either N, or { N }.

If this is not the case, name resolution does not resolve N when using min_const_generics as more complex expressions are not supported.

pub fn to_bound(&self) -> Option<GenericBound>

pub fn peel_parens(&self) -> &Expr

pub fn to_ty(&self) -> Option<P<Ty>>

Attempts to reparse as Ty (for diagnostic purposes).

pub fn precedence(&self) -> ExprPrecedence

pub fn take(&mut self) -> Self

pub fn is_approximately_pattern(&self) -> bool