rustc_ast::ast

Type Alias AssocItem

pub type AssocItem = Item<AssocItemKind>;

Represents associated items. These include items in impl and trait definitions.

Aliased Type

pub struct AssocItem {
    pub attrs: ThinVec<Attribute>,
    pub id: NodeId,
    pub span: Span,
    pub vis: Visibility,
    pub kind: AssocItemKind,
    pub tokens: Option<LazyAttrTokenStream>,
}

Fields
attrs: ThinVec<Attribute>id: NodeIdspan: Spanvis: Visibilitykind: AssocItemKindtokens: Option<LazyAttrTokenStream>

Original tokens this item was parsed from. This isn’t necessarily available for all items, although over time more and more items should have this be Some. Right now this is primarily used for procedural macros, notably custom attributes.

Note that the tokens here do not include the outer attributes, but will include inner attributes.