Type Alias rustc_ast::ast::ForeignItem

pub type ForeignItem = Item<ForeignItemKind>;

Aliased Type

struct ForeignItem {
    pub attrs: ThinVec<Attribute>,
    pub id: NodeId,
    pub span: Span,
    pub vis: Visibility,
    pub ident: Ident,
    pub kind: ForeignItemKind,
    pub tokens: Option<LazyAttrTokenStream>,
}

Fields
attrs: ThinVec<Attribute>id: NodeIdspan: Spanvis: Visibilityident: Ident

The name of the item. It might be a dummy name in case of anonymous items.

kind: ForeignItemKindtokens: Option<LazyAttrTokenStream>

Original tokens this item was parsed from. This isn’t necessarily available for all items, although over time more and more items should have this be Some. Right now this is primarily used for procedural macros, notably custom attributes.

Note that the tokens here do not include the outer attributes, but will include inner attributes.