Module rustc_ast::ast_traits

A set of traits implemented for various AST nodes, typically those used in AST fragments during macro expansion. The traits are not implemented exhaustively, only when actually necessary.

Macros

impl_has_attrs 🔒

impl_has_attrs_none 🔒

impl_has_node_id 🔒

impl_has_span 🔒

impl_has_tokens 🔒

impl_has_tokens_none 🔒

impl_not_ast_deref 🔒

Structs

AstNodeWrapper

A newtype around an AST node that implements the traits above if the node implements them.

Traits

AstDeref

A utility trait to reduce boilerplate. Standard Deref(Mut) cannot be reused due to coherence.

HasAttrs

A trait for AST nodes having (or not having) attributes.

HasNodeId

A trait for AST nodes having an ID.

HasSpan

A trait for AST nodes having a span.

HasTokens

A trait for AST nodes having (or not having) collected tokens.