Module rustc_ast::attr

Functions dealing with attributes and meta items.

Structs

AttrIdGenerator

MarkedAttrs

Statics

MAX_ATTR_ID 🔒

Functions

list_contains_name

mk_attr

mk_attr_from_item

mk_attr_inner

Returns an inner attribute with the given value and span.

mk_attr_outer

Returns an outer attribute with the given value and span.

mk_doc_comment

mk_list_item

mk_name_value_item

mk_name_value_item_str

mk_nested_word_item

mk_word_item