rustc_ast::mut_visit
Trait DummyAstNode

pub trait DummyAstNode {
    // Required method
    fn dummy() -> Self;
}

Some value for the AST node that is valid but possibly meaningless. Similar to Default but not intended for wide use. The value will never be used meaningfully, it exists just to support unwinding in visit_clobber in the case where its closure panics.

Required Methods

fn dummy() -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.