Trait rustc_ast::mut_visit::MutVisitor

pub trait MutVisitor: Sized {
    const VISIT_TOKENS: bool = false;

    fn visit_crate(&mut self, c: &mut Crate) { ... }
    fn visit_meta_list_item(&mut self, list_item: &mut NestedMetaItem) { ... }
    fn visit_meta_item(&mut self, meta_item: &mut MetaItem) { ... }
    fn visit_use_tree(&mut self, use_tree: &mut UseTree) { ... }
    fn flat_map_foreign_item(
        &mut self,
        ni: P<ForeignItem>
    ) -> SmallVec<[P<ForeignItem>; 1]> { ... }
    fn flat_map_item(&mut self, i: P<Item>) -> SmallVec<[P<Item>; 1]> { ... }
    fn visit_fn_header(&mut self, header: &mut FnHeader) { ... }
    fn flat_map_field_def(&mut self, fd: FieldDef) -> SmallVec<[FieldDef; 1]> { ... }
    fn visit_item_kind(&mut self, i: &mut ItemKind) { ... }
    fn flat_map_trait_item(
        &mut self,
        i: P<AssocItem>
    ) -> SmallVec<[P<AssocItem>; 1]> { ... }
    fn flat_map_impl_item(
        &mut self,
        i: P<AssocItem>
    ) -> SmallVec<[P<AssocItem>; 1]> { ... }
    fn visit_fn_decl(&mut self, d: &mut P<FnDecl>) { ... }
    fn visit_asyncness(&mut self, a: &mut Async) { ... }
    fn visit_closure_binder(&mut self, b: &mut ClosureBinder) { ... }
    fn visit_block(&mut self, b: &mut P<Block>) { ... }
    fn flat_map_stmt(&mut self, s: Stmt) -> SmallVec<[Stmt; 1]> { ... }
    fn flat_map_arm(&mut self, arm: Arm) -> SmallVec<[Arm; 1]> { ... }
    fn visit_pat(&mut self, p: &mut P<Pat>) { ... }
    fn visit_anon_const(&mut self, c: &mut AnonConst) { ... }
    fn visit_expr(&mut self, e: &mut P<Expr>) { ... }
    fn visit_method_receiver_expr(&mut self, ex: &mut P<Expr>) { ... }
    fn filter_map_expr(&mut self, e: P<Expr>) -> Option<P<Expr>> { ... }
    fn visit_generic_arg(&mut self, arg: &mut GenericArg) { ... }
    fn visit_ty(&mut self, t: &mut P<Ty>) { ... }
    fn visit_lifetime(&mut self, l: &mut Lifetime) { ... }
    fn visit_constraint(&mut self, t: &mut AssocConstraint) { ... }
    fn visit_foreign_mod(&mut self, nm: &mut ForeignMod) { ... }
    fn flat_map_variant(&mut self, v: Variant) -> SmallVec<[Variant; 1]> { ... }
    fn visit_ident(&mut self, i: &mut Ident) { ... }
    fn visit_path(&mut self, p: &mut Path) { ... }
    fn visit_qself(&mut self, qs: &mut Option<P<QSelf>>) { ... }
    fn visit_generic_args(&mut self, p: &mut GenericArgs) { ... }
    fn visit_angle_bracketed_parameter_data(
        &mut self,
        p: &mut AngleBracketedArgs
    ) { ... }
    fn visit_parenthesized_parameter_data(&mut self, p: &mut ParenthesizedArgs) { ... }
    fn visit_local(&mut self, l: &mut P<Local>) { ... }
    fn visit_mac_call(&mut self, mac: &mut MacCall) { ... }
    fn visit_macro_def(&mut self, def: &mut MacroDef) { ... }
    fn visit_label(&mut self, label: &mut Label) { ... }
    fn visit_attribute(&mut self, at: &mut Attribute) { ... }
    fn flat_map_param(&mut self, param: Param) -> SmallVec<[Param; 1]> { ... }
    fn visit_generics(&mut self, generics: &mut Generics) { ... }
    fn visit_trait_ref(&mut self, tr: &mut TraitRef) { ... }
    fn visit_poly_trait_ref(&mut self, p: &mut PolyTraitRef) { ... }
    fn visit_variant_data(&mut self, vdata: &mut VariantData) { ... }
    fn flat_map_generic_param(
        &mut self,
        param: GenericParam
    ) -> SmallVec<[GenericParam; 1]> { ... }
    fn visit_param_bound(&mut self, tpb: &mut GenericBound) { ... }
    fn visit_mt(&mut self, mt: &mut MutTy) { ... }
    fn flat_map_expr_field(&mut self, f: ExprField) -> SmallVec<[ExprField; 1]> { ... }
    fn visit_where_clause(&mut self, where_clause: &mut WhereClause) { ... }
    fn visit_where_predicate(&mut self, where_predicate: &mut WherePredicate) { ... }
    fn visit_vis(&mut self, vis: &mut Visibility) { ... }
    fn visit_id(&mut self, _id: &mut NodeId) { ... }
    fn visit_span(&mut self, _sp: &mut Span) { ... }
    fn flat_map_pat_field(&mut self, fp: PatField) -> SmallVec<[PatField; 1]> { ... }
    fn visit_inline_asm(&mut self, asm: &mut InlineAsm) { ... }
    fn visit_inline_asm_sym(&mut self, sym: &mut InlineAsmSym) { ... }

}

Provided Associated Constants

const VISIT_TOKENS: bool = false

Mutable token visiting only exists for the macro_rules token marker and should not be used otherwise. Token visitor would be entirely separate from the regular visitor if the marker didn’t have to visit AST fragments in nonterminal tokens.

Provided Methods

fn visit_crate(&mut self, c: &mut Crate)

fn visit_meta_list_item(&mut self, list_item: &mut NestedMetaItem)

fn visit_meta_item(&mut self, meta_item: &mut MetaItem)

fn visit_use_tree(&mut self, use_tree: &mut UseTree)

fn flat_map_foreign_item(
&mut self,
ni: P<ForeignItem>
) -> SmallVec<[P<ForeignItem>; 1]>

fn flat_map_item(&mut self, i: P<Item>) -> SmallVec<[P<Item>; 1]>

fn visit_fn_header(&mut self, header: &mut FnHeader)

fn flat_map_field_def(&mut self, fd: FieldDef) -> SmallVec<[FieldDef; 1]>

fn visit_item_kind(&mut self, i: &mut ItemKind)

fn flat_map_trait_item(&mut self, i: P<AssocItem>) -> SmallVec<[P<AssocItem>; 1]>

fn flat_map_impl_item(&mut self, i: P<AssocItem>) -> SmallVec<[P<AssocItem>; 1]>

fn visit_fn_decl(&mut self, d: &mut P<FnDecl>)

fn visit_asyncness(&mut self, a: &mut Async)

fn visit_closure_binder(&mut self, b: &mut ClosureBinder)

fn visit_block(&mut self, b: &mut P<Block>)

fn flat_map_stmt(&mut self, s: Stmt) -> SmallVec<[Stmt; 1]>

fn flat_map_arm(&mut self, arm: Arm) -> SmallVec<[Arm; 1]>

fn visit_pat(&mut self, p: &mut P<Pat>)

fn visit_anon_const(&mut self, c: &mut AnonConst)

fn visit_expr(&mut self, e: &mut P<Expr>)

fn visit_method_receiver_expr(&mut self, ex: &mut P<Expr>)

This method is a hack to workaround unstable of stmt_expr_attributes. It can be removed once that feature is stabilized.

fn filter_map_expr(&mut self, e: P<Expr>) -> Option<P<Expr>>

fn visit_generic_arg(&mut self, arg: &mut GenericArg)

fn visit_ty(&mut self, t: &mut P<Ty>)

fn visit_lifetime(&mut self, l: &mut Lifetime)

fn visit_constraint(&mut self, t: &mut AssocConstraint)

fn visit_foreign_mod(&mut self, nm: &mut ForeignMod)

fn flat_map_variant(&mut self, v: Variant) -> SmallVec<[Variant; 1]>

fn visit_ident(&mut self, i: &mut Ident)

fn visit_path(&mut self, p: &mut Path)

fn visit_qself(&mut self, qs: &mut Option<P<QSelf>>)

fn visit_generic_args(&mut self, p: &mut GenericArgs)

fn visit_angle_bracketed_parameter_data(&mut self, p: &mut AngleBracketedArgs)

fn visit_parenthesized_parameter_data(&mut self, p: &mut ParenthesizedArgs)

fn visit_local(&mut self, l: &mut P<Local>)

fn visit_mac_call(&mut self, mac: &mut MacCall)

fn visit_macro_def(&mut self, def: &mut MacroDef)

fn visit_label(&mut self, label: &mut Label)

fn visit_attribute(&mut self, at: &mut Attribute)

fn flat_map_param(&mut self, param: Param) -> SmallVec<[Param; 1]>

fn visit_generics(&mut self, generics: &mut Generics)

fn visit_trait_ref(&mut self, tr: &mut TraitRef)

fn visit_poly_trait_ref(&mut self, p: &mut PolyTraitRef)

fn visit_variant_data(&mut self, vdata: &mut VariantData)

fn flat_map_generic_param(
&mut self,
param: GenericParam
) -> SmallVec<[GenericParam; 1]>

fn visit_param_bound(&mut self, tpb: &mut GenericBound)

fn visit_mt(&mut self, mt: &mut MutTy)

fn flat_map_expr_field(&mut self, f: ExprField) -> SmallVec<[ExprField; 1]>

fn visit_where_clause(&mut self, where_clause: &mut WhereClause)

fn visit_where_predicate(&mut self, where_predicate: &mut WherePredicate)

fn visit_vis(&mut self, vis: &mut Visibility)

fn visit_id(&mut self, _id: &mut NodeId)

fn visit_span(&mut self, _sp: &mut Span)

fn flat_map_pat_field(&mut self, fp: PatField) -> SmallVec<[PatField; 1]>

fn visit_inline_asm(&mut self, asm: &mut InlineAsm)

fn visit_inline_asm_sym(&mut self, sym: &mut InlineAsmSym)