rustc_ast::ptr

Type Alias P

pub type P<T> = Box<T>;

A pointer type that uniquely owns a heap allocation of type T.

This used to be its own type, but now it’s just a typedef for Box and we are planning to remove it soon.

Aliased Type

pub struct P<T>(/* private fields */);