Module rustc_ast::token

Re-exports

pub use BinOpToken::*;

pub use LitKind::*;

pub use Nonterminal::*;

pub use TokenKind::*;

Structs

Lit

A literal token.

Token

Enums

BinOpToken

CommentKind

Delimiter

Describes how a sequence of token trees is delimited. Cannot use proc_macro::Delimiter directly because this structure should implement some additional traits. The None variant is also renamed to Invisible to be less confusing and better convey the semantics.

LitKind

Nonterminal

For interpolation during macro expansion.

NonterminalKind

TokenKind

Functions

ident_can_begin_expr

ident_can_begin_type 🔒