Struct rustc_ast::tokenstream::TokenStream

pub struct TokenStream(pub(crate) Lrc<Vec<TokenTree>>);

A TokenStream is an abstract sequence of tokens, organized into TokenTrees.

The goal is for procedural macros to work with TokenStreams and TokenTrees instead of a representation of the abstract syntax tree. Todayโs TokenTrees can still contain AST via token::Interpolated for backwards compatibility.

Tuple Fields
0: Lrc<Vec<TokenTree>>
Implementations

impl TokenStream

pub fn add_comma(&self) -> Option<(TokenStream, Span)>

Given a TokenStream with a Stream of only two arguments, return a new TokenStream separating the two arguments with a comma for diagnostic suggestions.

impl TokenStream

pub fn new(tts: Vec<TokenTree>) -> TokenStream

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

pub fn trees(&self) -> RefTokenTreeCursor<'_>

pub fn into_trees(self) -> TokenTreeCursor

pub fn eq_unspanned(&self, other: &TokenStream) -> bool

Compares two TokenStreams, checking equality without regarding span information.

pub fn token_alone(kind: TokenKind, span: Span) -> TokenStream

Create a token stream containing a single token with alone spacing. The spacing used for the final token in a constructed stream doesnโt matter because itโs never used. In practice we arbitrarily use Spacing::Alone.

pub fn from_ast(node: &(impl HasAttrs + HasTokens + Debug)) -> TokenStream

pub fn from_nonterminal_ast(nt: &Nonterminal) -> TokenStream

fn flatten_token(token: &Token, spacing: Spacing) -> TokenTree

fn flatten_token_tree(tree: &TokenTree) -> TokenTree

pub fn flattened(&self) -> TokenStream

fn try_glue_to_last(vec: &mut Vec<TokenTree>, tt: &TokenTree) -> bool

pub fn push_tree(&mut self, tt: TokenTree)

Push tt onto the end of the stream, possibly gluing it to the last token. Uses make_mut to maximize efficiency.

pub fn push_stream(&mut self, stream: TokenStream)

Push stream onto the end of the stream, possibly gluing the first token tree to the last token. (No other token trees will be glued.) Uses make_mut to maximize efficiency.

pub fn chunks(&self, chunk_size: usize) -> Chunks<'_, TokenTree>

pub fn desugar_doc_comments(&mut self)

Desugar doc comments like /// foo in the stream into #[doc = r"foo"]. Modifies the TokenStream via Lrc::make_mut, but as little as possible.