rustc_ast
Trait HashStableContext

pub trait HashStableContext: HashStableContext {
    // Required method
    fn hash_attr(&mut self, _: &Attribute, hasher: &mut StableHasher);
}

Requirements for a StableHashingContext to be used in this crate. This is a hack to allow using the HashStable_Generic derive macro instead of implementing everything in rustc_middle.

Required Methods

fn hash_attr(&mut self, _: &Attribute, hasher: &mut StableHasher)