Module rustc_ast::util::comments

Structs

Comment

Enums

CommentStyle

Functions

all_whitespace 🔒

Returns None if the first col chars of s contain a non-whitespace char. Otherwise returns Some(k) where k is first char offset after that leading whitespace. Note that k may be outside bounds of s.

beautify_doc_string

Makes a doc string more presentable to users. Used by rustdoc and perhaps other tools, but not by rustc.

gather_comments

may_have_doc_links

A fast conservative estimate on whether the string can contain documentation links. A pair of square brackets [] must exist in the string, but we only search for the opening bracket because brackets always go in pairs in practice.

split_block_comment_into_lines 🔒

trim_whitespace_prefix 🔒