Enum rustc_ast::ast::ExprKind

pub enum ExprKind {

    Array(ThinVec<P<Expr>>),
    ConstBlock(AnonConst),
    Call(P<Expr>, ThinVec<P<Expr>>),
    MethodCall(Box<MethodCall>),
    Tup(ThinVec<P<Expr>>),
    Binary(BinOp, P<Expr>, P<Expr>),
    Unary(UnOp, P<Expr>),
    Lit(Lit),
    Cast(P<Expr>, P<Ty>),
    Type(P<Expr>, P<Ty>),
    Let(P<Pat>, P<Expr>, Span, Option<ErrorGuaranteed>),
    If(P<Expr>, P<Block>, Option<P<Expr>>),
    While(P<Expr>, P<Block>, Option<Label>),
    ForLoop {
        pat: P<Pat>,
        iter: P<Expr>,
        body: P<Block>,
        label: Option<Label>,
        kind: ForLoopKind,
    },
    Loop(P<Block>, Option<Label>, Span),
    Match(P<Expr>, ThinVec<Arm>),
    Closure(Box<Closure>),
    Block(P<Block>, Option<Label>),
    Gen(CaptureBy, P<Block>, GenBlockKind),
    Await(P<Expr>, Span),
    TryBlock(P<Block>),
    Assign(P<Expr>, P<Expr>, Span),
    AssignOp(BinOp, P<Expr>, P<Expr>),
    Field(P<Expr>, Ident),
    Index(P<Expr>, P<Expr>, Span),
    Range(Option<P<Expr>>, Option<P<Expr>>, RangeLimits),
    Underscore,
    Path(Option<P<QSelf>>, Path),
    AddrOf(BorrowKind, Mutability, P<Expr>),
    Break(Option<Label>, Option<P<Expr>>),
    Continue(Option<Label>),
    Ret(Option<P<Expr>>),
    InlineAsm(P<InlineAsm>),
    OffsetOf(P<Ty>, P<[Ident]>),
    MacCall(P<MacCall>),
    Struct(P<StructExpr>),
    Repeat(P<Expr>, AnonConst),
    Paren(P<Expr>),
    Try(P<Expr>),
    Yield(Option<P<Expr>>),
    Yeet(Option<P<Expr>>),
    Become(P<Expr>),
    IncludedBytes(Lrc<[u8]>),
    FormatArgs(P<FormatArgs>),
    Err(ErrorGuaranteed),
    Dummy,

}

Variants

Array(ThinVec<P<Expr>>)

An array ([a, b, c, d])

ConstBlock(AnonConst)

Allow anonymous constants from an inline const block

Call(P<Expr>, ThinVec<P<Expr>>)

A function call

The first field resolves to the function itself, and the second field is the list of arguments. This also represents calling the constructor of tuple-like ADTs such as tuple structs and enum variants.

MethodCall(Box<MethodCall>)

A method call (e.g. x.foo::<Bar, Baz>(a, b, c)).

Tup(ThinVec<P<Expr>>)

A tuple (e.g., (a, b, c, d)).

Binary(BinOp, P<Expr>, P<Expr>)

A binary operation (e.g., a + b, a * b).

Unary(UnOp, P<Expr>)

A unary operation (e.g., !x, *x).

Lit(Lit)

A literal (e.g., 1, "foo").

Cast(P<Expr>, P<Ty>)

A cast (e.g., foo as f64).

Type(P<Expr>, P<Ty>)

A type ascription (e.g., 42: usize).

Let(P<Pat>, P<Expr>, Span, Option<ErrorGuaranteed>)

A let pat = expr expression that is only semantically allowed in the condition of if / while expressions. (e.g., if let 0 = x { .. }).

Span represents the whole let pat = expr statement.

If(P<Expr>, P<Block>, Option<P<Expr>>)

An if block, with an optional else block.

if expr { block } else { expr }

While(P<Expr>, P<Block>, Option<Label>)

A while loop, with an optional label.

'label: while expr { block }

ForLoop

Fields

pat: P<Pat>

iter: P<Expr>

body: P<Block>

label: Option<Label>

kind: ForLoopKind

A for loop, with an optional label.

'label: for await? pat in iter { block }

This is desugared to a combination of loop and match expressions.

Loop(P<Block>, Option<Label>, Span)

Conditionless loop (can be exited with break, continue, or return).

'label: loop { block }

Match(P<Expr>, ThinVec<Arm>)

A match block.

Closure(Box<Closure>)

A closure (e.g., move |a, b, c| a + b + c).

Block(P<Block>, Option<Label>)

A block ('label: { ... }).

Gen(CaptureBy, P<Block>, GenBlockKind)

An async block (async move { ... }), or a gen block (gen move { ... })

Await(P<Expr>, Span)

An await expression (my_future.await). Span is of await keyword.

TryBlock(P<Block>)

A try block (try { ... }).

Assign(P<Expr>, P<Expr>, Span)

An assignment (a = foo()). The Span argument is the span of the = token.

AssignOp(BinOp, P<Expr>, P<Expr>)

An assignment with an operator.

E.g., a += 1.

Field(P<Expr>, Ident)

Access of a named (e.g., obj.foo) or unnamed (e.g., obj.0) struct field.

Index(P<Expr>, P<Expr>, Span)

An indexing operation (e.g., foo[2]). The span represents the span of the [2], including brackets.

Range(Option<P<Expr>>, Option<P<Expr>>, RangeLimits)

A range (e.g., 1..2, 1.., ..2, 1..=2, ..=2; and .. in destructuring assignment).

Underscore

An underscore, used in destructuring assignment to ignore a value.

Path(Option<P<QSelf>>, Path)

Variable reference, possibly containing :: and/or type parameters (e.g., foo::bar::<baz>).

Optionally “qualified” (e.g., <Vec<T> as SomeTrait>::SomeType).

AddrOf(BorrowKind, Mutability, P<Expr>)

A referencing operation (&a, &mut a, &raw const a or &raw mut a).

Break(Option<Label>, Option<P<Expr>>)

A break, with an optional label to break, and an optional expression.

Continue(Option<Label>)

A continue, with an optional label.

Ret(Option<P<Expr>>)

A return, with an optional value to be returned.

InlineAsm(P<InlineAsm>)

Output of the asm!() macro.

OffsetOf(P<Ty>, P<[Ident]>)

Output of the offset_of!() macro.

MacCall(P<MacCall>)

A macro invocation; pre-expansion.

Struct(P<StructExpr>)

A struct literal expression.

E.g., Foo {x: 1, y: 2}, or Foo {x: 1, .. rest}.

Repeat(P<Expr>, AnonConst)

An array literal constructed from one repeated element.

E.g., [1; 5]. The expression is the element to be repeated; the constant is the number of times to repeat it.

Paren(P<Expr>)

No-op: used solely so we can pretty-print faithfully.

Try(P<Expr>)

A try expression (expr?).

Yield(Option<P<Expr>>)

A yield, with an optional value to be yielded.

Yeet(Option<P<Expr>>)

A do yeet (aka throw/fail/bail/raise/whatever), with an optional value to be returned.

Become(P<Expr>)

A tail call return, with the value to be returned.

While .0 must be a function call, we check this later, after parsing.

IncludedBytes(Lrc<[u8]>)

Bytes included via include_bytes! Added for optimization purposes to avoid the need to escape large binary blobs - should always behave like ExprKind::Lit with a ByteStr literal.

FormatArgs(P<FormatArgs>)

A format_args!() expression.

Err(ErrorGuaranteed)

Placeholder for an expression that wasn’t syntactically well formed in some way.

Dummy

Acts as a null expression. Lowering it will always emit a bug.