Enum rustc_ast::ast::ItemKind

pub enum ItemKind {

    ExternCrate(Option<Symbol>),
    Use(UseTree),
    Static(Box<StaticItem>),
    Const(Box<ConstItem>),
    Fn(Box<Fn>),
    Mod(Unsafe, ModKind),
    ForeignMod(ForeignMod),
    GlobalAsm(Box<InlineAsm>),
    TyAlias(Box<TyAlias>),
    Enum(EnumDef, Generics),
    Struct(VariantData, Generics),
    Union(VariantData, Generics),
    Trait(Box<Trait>),
    TraitAlias(Generics, GenericBounds),
    Impl(Box<Impl>),
    MacCall(P<MacCall>),
    MacroDef(MacroDef),

}

Variants

ExternCrate(Option<Symbol>)

An extern crate item, with the optional original crate name if the crate was renamed.

E.g., extern crate foo or extern crate foo_bar as foo.

Use(UseTree)

A use declaration item (use).

E.g., use foo;, use foo::bar; or use foo::bar as FooBar;.

Static(Box<StaticItem>)

A static item (static).

E.g., static FOO: i32 = 42; or static FOO: &'static str = "bar";.

Const(Box<ConstItem>)

A constant item (const).

E.g., const FOO: i32 = 42;.

Fn(Box<Fn>)

A function declaration (fn).

E.g., fn foo(bar: usize) -> usize { .. }.

Mod(Unsafe, ModKind)

A module declaration (mod).

E.g., mod foo; or mod foo { .. }. unsafe keyword on modules is accepted syntactically for macro DSLs, but not semantically by Rust.

ForeignMod(ForeignMod)

An external module (extern).

E.g., extern {} or extern "C" {}.

GlobalAsm(Box<InlineAsm>)

Module-level inline assembly (from global_asm!()).

TyAlias(Box<TyAlias>)

A type alias (type).

E.g., type Foo = Bar<u8>;.

Enum(EnumDef, Generics)

An enum definition (enum).

E.g., enum Foo<A, B> { C<A>, D<B> }.

Struct(VariantData, Generics)

A struct definition (struct).

E.g., struct Foo<A> { x: A }.

Union(VariantData, Generics)

A union definition (union).

E.g., union Foo<A, B> { x: A, y: B }.

Trait(Box<Trait>)

A trait declaration (trait).

E.g., trait Foo { .. }, trait Foo<T> { .. } or auto trait Foo {}.

TraitAlias(Generics, GenericBounds)

Trait alias

E.g., trait Foo = Bar + Quux;.

Impl(Box<Impl>)

An implementation.

E.g., impl<A> Foo<A> { .. } or impl<A> Trait for Foo<A> { .. }.

MacCall(P<MacCall>)

A macro invocation.

E.g., foo!(..).

MacroDef(MacroDef)

A macro definition.

Implementations

impl ItemKind

pub fn article(&self) -> &'static str

pub fn descr(&self) -> &'static str

pub fn generics(&self) -> Option<&Generics>