Enum rustc_ast::ast::LitKind

pub enum LitKind {
    Str(Symbol, StrStyle),
    ByteStr(Lrc<[u8]>, StrStyle),
    CStr(Lrc<[u8]>, StrStyle),
    Byte(u8),
    Char(char),
    Int(Pu128, LitIntType),
    Float(Symbol, LitFloatType),
    Bool(bool),
    Err(ErrorGuaranteed),
}

This type is used within both ast::MetaItemLit and hir::Lit.

Note that the entire literal (including the suffix) is considered when deciding the LitKind. This means that float literals like 1f32 are classified by this type as Float. This is different to token::LitKind which does not consider the suffix.

Variants

Str(Symbol, StrStyle)

A string literal ("foo"). The symbol is unescaped, and so may differ from the original token’s symbol.

ByteStr(Lrc<[u8]>, StrStyle)

A byte string (b"foo"). Not stored as a symbol because it might be non-utf8, and symbols only allow utf8 strings.

CStr(Lrc<[u8]>, StrStyle)

A C String (c"foo"). Guaranteed to only have \0 at the end.

Byte(u8)

A byte char (b'f').

Char(char)

A character literal ('a').

Int(Pu128, LitIntType)

An integer literal (1).

Float(Symbol, LitFloatType)

A float literal (1.0, 1f64 or 1E10f64). The pre-suffix part is stored as a symbol rather than f64 so that LitKind can impl Eq and Hash.

Bool(bool)

A boolean literal (true, false).

Err(ErrorGuaranteed)

Placeholder for a literal that wasn’t well-formed in some way.

Implementations

impl LitKind

pub fn from_token_lit(lit: Lit) -> Result<LitKind, LitError>

Converts literal token into a semantic literal.

impl LitKind

pub fn str(&self) -> Option<Symbol>

pub fn is_str(&self) -> bool

Returns true if this literal is a string.

pub fn is_bytestr(&self) -> bool

Returns true if this literal is byte literal string.

pub fn is_numeric(&self) -> bool

Returns true if this is a numeric literal.

pub fn is_unsuffixed(&self) -> bool

Returns true if this literal has no suffix. Note: this will return true for literals with prefixes such as raw strings and byte strings.

pub fn is_suffixed(&self) -> bool

Returns true if this literal has a suffix.