Enum rustc_ast::ast::TyKind

pub enum TyKind {

    Slice(P<Ty>),
    Array(P<Ty>, AnonConst),
    Ptr(MutTy),
    Ref(Option<Lifetime>, MutTy),
    BareFn(P<BareFnTy>),
    Never,
    Tup(ThinVec<P<Ty>>),
    AnonStruct(ThinVec<FieldDef>),
    AnonUnion(ThinVec<FieldDef>),
    Path(Option<P<QSelf>>, Path),
    TraitObject(GenericBounds, TraitObjectSyntax),
    ImplTrait(NodeId, GenericBounds),
    Paren(P<Ty>),
    Typeof(AnonConst),
    Infer,
    ImplicitSelf,
    MacCall(P<MacCall>),
    Err,
    CVarArgs,

}

The various kinds of type recognized by the compiler.

Variants

Slice(P<Ty>)

A variable-length slice ([T]).

Array(P<Ty>, AnonConst)

A fixed length array ([T; n]).

Ptr(MutTy)

A raw pointer (*const T or *mut T).

Ref(Option<Lifetime>, MutTy)

A reference (&'a T or &'a mut T).

BareFn(P<BareFnTy>)

A bare function (e.g., fn(usize) -> bool).

Never

The never type (!).

Tup(ThinVec<P<Ty>>)

A tuple ((A, B, C, D,...)).

AnonStruct(ThinVec<FieldDef>)

An anonymous struct type i.e. struct { foo: Type }

AnonUnion(ThinVec<FieldDef>)

An anonymous union type i.e. union { bar: Type }

Path(Option<P<QSelf>>, Path)

A path (module::module::...::Type), optionally “qualified”, e.g., <Vec<T> as SomeTrait>::SomeType.

Type parameters are stored in the Path itself.

TraitObject(GenericBounds, TraitObjectSyntax)

A trait object type Bound1 + Bound2 + Bound3 where Bound is a trait or a lifetime.

ImplTrait(NodeId, GenericBounds)

An impl Bound1 + Bound2 + Bound3 type where Bound is a trait or a lifetime.

The NodeId exists to prevent lowering from having to generate NodeIds on the fly, which would complicate the generation of opaque type Foo = impl Trait items significantly.

Paren(P<Ty>)

No-op; kept solely so that we can pretty-print faithfully.

Typeof(AnonConst)

Unused for now.

Infer

This means the type should be inferred instead of it having been specified. This can appear anywhere in a type.

ImplicitSelf

Inferred type of a self or &self argument in a method.

MacCall(P<MacCall>)

A macro in the type position.

Err

Placeholder for a kind that has failed to be defined.

CVarArgs

Placeholder for a va_list.

Implementations

impl TyKind

pub fn is_implicit_self(&self) -> bool

pub fn is_unit(&self) -> bool

pub fn is_simple_path(&self) -> Option<Symbol>