Module rustc_ast::mut_visit

A MutVisitor represents an AST modification; it accepts an AST piece and mutates it in place. So, for instance, macro expansion is a MutVisitor that walks over an AST and modifies it.

Note: using a MutVisitor (other than the MacroExpander MutVisitor) on an AST before macro expansion is probably a bad idea. For instance, a MutVisitor renaming item names in a module will miss all of those that are created by the expansion of a macro.

Traits

DummyAstNode

Some value for the AST node that is valid but possibly meaningless. Similar to Default but not intended for wide use. The value will never be used meaningfully, it exists just to support unwinding in visit_clobber in the case where its closure panics.

ExpectOne

MutVisitor

Functions

noop_filter_map_expr

noop_flat_map_arm

noop_flat_map_assoc_item

noop_flat_map_expr_field

noop_flat_map_field_def

noop_flat_map_foreign_item

noop_flat_map_generic_param

noop_flat_map_item

noop_flat_map_param

noop_flat_map_pat_field

noop_flat_map_stmt

noop_flat_map_stmt_kind

noop_flat_map_variant

noop_visit_angle_bracketed_parameter_data

noop_visit_anon_const

noop_visit_attribute

noop_visit_block

noop_visit_capture_by

noop_visit_closure_binder

noop_visit_constraint

noop_visit_coroutine_kind

noop_visit_crate

noop_visit_expr

noop_visit_fn_decl

noop_visit_fn_header

noop_visit_fn_ret_ty

noop_visit_foreign_mod

noop_visit_format_args

noop_visit_generic_arg

noop_visit_generic_args

noop_visit_generics

noop_visit_ident

noop_visit_inline_asm

noop_visit_inline_asm_sym

noop_visit_item_kind

noop_visit_label

noop_visit_lifetime 🔒

noop_visit_local

noop_visit_mac

noop_visit_macro_def

noop_visit_meta_item

noop_visit_meta_list_item

noop_visit_mt

noop_visit_param_bound

noop_visit_parenthesized_parameter_data

noop_visit_pat

noop_visit_path

noop_visit_poly_trait_ref

noop_visit_qself

noop_visit_trait_ref

noop_visit_ty

noop_visit_use_tree

noop_visit_variant_data

noop_visit_vis

noop_visit_where_clause

noop_visit_where_predicate

visit_attr_args

visit_attr_tt

visit_attr_tts

visit_attrs

visit_bounds

visit_clobber

Use a map-style function (FnOnce(T) -> T) to overwrite a &mut T. Useful when using a flat_map_* or filter_map_* method within a visit_ method.

visit_const_item 🔒

visit_constness

visit_defaultness

visit_delim_args

visit_delim_span

visit_exprs

visit_fn_sig

visit_lazy_tts

visit_lazy_tts_opt_mut

visit_nonterminal

Applies the visitor to elements of interpolated nodes.

visit_opt

visit_polarity

visit_thin_exprs

visit_thin_vec

visit_token

Applies ident visitor if it’s an ident; applies other visits to interpolated nodes. In practice the ident part is not actually used by specific visitors right now, but there’s a test below checking that it works.

visit_tt

visit_tts

visit_unsafety

visit_vec