Trait rustc_ast::mut_visit::DummyAstNode

pub trait DummyAstNode {
    // Required method
    fn dummy() -> Self;
}

Some value for the AST node that is valid but possibly meaningless. Similar to Default but not intended for wide use. The value will never be used meaningfully, it exists just to support unwinding in visit_clobber in the case where its closure panics.

Required Methods

fn dummy() -> Self

Object Safety

This trait is not object safe.