rustc_ast
Module token

Re-exports

pub use BinOpToken::*;

pub use LitKind::*;

pub use Nonterminal::*;

pub use NtExprKind::*;

pub use NtPatKind::*;

pub use TokenKind::*;

Modules

size_asserts 🔒

Structs

Lit

A literal token.

Token

Enums

BinOpToken

CommentKind

Delimiter

Describes how a sequence of token trees is delimited. Cannot use proc_macro::Delimiter directly because this structure should implement some additional traits.

IdentIsRaw

InvisibleOrigin

LitKind

MetaVarKind

Annoyingly similar to NonterminalKind, but the slight differences are important.

Nonterminal

For interpolation during macro expansion.

NonterminalKind

NtExprKind

NtPatKind

TokenKind

Functions

ident_can_begin_expr

ident_can_begin_type 🔒