rustc_ast

Module token

Re-exports

pub use LitKind::*;

pub use NtExprKind::*;

pub use NtPatKind::*;

pub use TokenKind::*;

Modules

size_asserts 🔒

Structs

Lit

A literal token.

Token

Enums

CommentKind

Delimiter

Describes how a sequence of token trees is delimited. Cannot use proc_macro::Delimiter directly because this structure should implement some additional traits.

IdentIsRaw

InvisibleOrigin

LitKind

MetaVarKind

Annoyingly similar to NonterminalKind, but the slight differences are important.

NonterminalKind

A macro nonterminal, known in documentation as a fragment specifier.

NtExprKind

NtPatKind

TokenKind

Functions

ident_can_begin_expr

ident_can_begin_type 🔒