Enum rustc_ast::token::TokenKind

pub enum TokenKind {

    Eq,
    Lt,
    Le,
    EqEq,
    Ne,
    Ge,
    Gt,
    AndAnd,
    OrOr,
    Not,
    Tilde,
    BinOp(BinOpToken),
    BinOpEq(BinOpToken),
    At,
    Dot,
    DotDot,
    DotDotDot,
    DotDotEq,
    Comma,
    Semi,
    Colon,
    PathSep,
    RArrow,
    LArrow,
    FatArrow,
    Pound,
    Dollar,
    Question,
    SingleQuote,
    OpenDelim(Delimiter),
    CloseDelim(Delimiter),
    Literal(Lit),
    Ident(Symbol, IdentIsRaw),
    NtIdent(Ident, IdentIsRaw),
    Lifetime(Symbol),
    NtLifetime(Ident),
    Interpolated(Lrc<Nonterminal>),
    DocComment(CommentKind, AttrStyle, Symbol),
    Eof,

}

Variants

Eq

=

Lt

<

Le

<=

EqEq

==

Ne

!=

Ge

>=

Gt

>

AndAnd

&&

OrOr

||

Not

!

Tilde

~

BinOp(BinOpToken)

BinOpEq(BinOpToken)

At

@

Dot

.

DotDot

..

DotDotDot

...

DotDotEq

..=

Comma

,

Semi

;

Colon

:

PathSep

::

RArrow

->

LArrow

<-

FatArrow

=>

Pound

#

Dollar

$

Question

?

SingleQuote

Used by proc macros for representing lifetimes, not generated by lexer right now.

OpenDelim(Delimiter)

An opening delimiter (e.g., {).

CloseDelim(Delimiter)

A closing delimiter (e.g., }).

Literal(Lit)

Ident(Symbol, IdentIsRaw)

Identifier token. Do not forget about NtIdent when you want to match on identifiers. Itโs recommended to use Token::(ident,uninterpolate,uninterpolated_span) to treat regular and interpolated identifiers in the same way.

NtIdent(Ident, IdentIsRaw)

This identifier (and its span) is the identifier passed to the declarative macro. The span in the surrounding Token is the span of the ident metavariable in the macroโs RHS.

Lifetime(Symbol)

Lifetime identifier token. Do not forget about NtLifetime when you want to match on lifetime identifiers. Itโs recommended to use Token::(lifetime,uninterpolate,uninterpolated_span) to treat regular and interpolated lifetime identifiers in the same way.

NtLifetime(Ident)

This identifier (and its span) is the lifetime passed to the declarative macro. The span in the surrounding Token is the span of the lifetime metavariable in the macroโs RHS.

Interpolated(Lrc<Nonterminal>)

An embedded AST node, as produced by a macro. This only exists for historical reasons. Weโd like to get rid of it, for multiple reasons.

Itโs conceptually very strange. Saying a token can contain an AST node is like saying, in natural language, that a word can contain a sentence.

It requires special handling in a bunch of places in the parser.

It prevents Token from implementing Copy. It adds complexity and likely slows things down. Please donโt add new occurrences of this token kind!

The span in the surrounding Token is that of the metavariable in the macroโs RHS. The span within the Nonterminal is that of the fragment passed to the macro at the call site.

DocComment(CommentKind, AttrStyle, Symbol)

A doc comment token. Symbol is the doc commentโs data excluding its โquotesโ (///, /**, etc) similarly to symbols in string literal tokens.

Eof

End Of File

Implementations

impl TokenKind

pub fn lit(kind: LitKind, symbol: Symbol, suffix: Option<Symbol>) -> TokenKind

pub fn break_two_token_op(&self) -> Option<(TokenKind, TokenKind)>

An approximation to proc-macro-style single-character operators used by rustc parser. If the operator token can be broken into two tokens, the first of which is single-character, then this function performs that operation, otherwise it returns None.

pub fn similar_tokens(&self) -> Option<Vec<TokenKind>>

Returns tokens that are likely to be typed accidentally instead of the current token. Enables better error recovery when the wrong token is found.

pub fn should_end_const_arg(&self) -> bool