rustc_ast::token

Enum TokenKind

pub enum TokenKind {

    Eq,
    Lt,
    Le,
    EqEq,
    Ne,
    Ge,
    Gt,
    AndAnd,
    OrOr,
    Bang,
    Tilde,
    Plus,
    Minus,
    Star,
    Slash,
    Percent,
    Caret,
    And,
    Or,
    Shl,
    Shr,
    PlusEq,
    MinusEq,
    StarEq,
    SlashEq,
    PercentEq,
    CaretEq,
    AndEq,
    OrEq,
    ShlEq,
    ShrEq,
    At,
    Dot,
    DotDot,
    DotDotDot,
    DotDotEq,
    Comma,
    Semi,
    Colon,
    PathSep,
    RArrow,
    LArrow,
    FatArrow,
    Pound,
    Dollar,
    Question,
    SingleQuote,
    OpenParen,
    CloseParen,
    OpenBrace,
    CloseBrace,
    OpenBracket,
    CloseBracket,
    OpenInvisible(InvisibleOrigin),
    CloseInvisible(InvisibleOrigin),
    Literal(Lit),
    Ident(Symbol, IdentIsRaw),
    NtIdent(Ident, IdentIsRaw),
    Lifetime(Symbol, IdentIsRaw),
    NtLifetime(Ident, IdentIsRaw),
    DocComment(CommentKind, AttrStyle, Symbol),
    Eof,

}

Variants

Eq

=

Lt

<

Le

<=

EqEq

==

Ne

!=

Ge

>=

Gt

>

AndAnd

&&

OrOr

||

Bang

!

Tilde

~

Plus

Minus

Star

Slash

Percent

Caret

And

Or

Shl

Shr

PlusEq

MinusEq

StarEq

SlashEq

PercentEq

CaretEq

AndEq

OrEq

ShlEq

ShrEq

At

@

Dot

.

DotDot

..

DotDotDot

...

DotDotEq

..=

Comma

,

Semi

;

Colon

:

PathSep

::

RArrow

->

LArrow

<-

FatArrow

=>

Pound

#

Dollar

$

Question

?

SingleQuote

Used by proc macros for representing lifetimes, not generated by lexer right now.

OpenParen

(

CloseParen

)

OpenBrace

{

CloseBrace

}

OpenBracket

[

CloseBracket

]

OpenInvisible(InvisibleOrigin)

Invisible opening delimiter, produced by a macro.

CloseInvisible(InvisibleOrigin)

Invisible closing delimiter, produced by a macro.

Literal(Lit)

Ident(Symbol, IdentIsRaw)

Identifier token. Do not forget about NtIdent when you want to match on identifiers. It’s recommended to use Token::{ident,uninterpolate} and Parser::token_uninterpolated_span to treat regular and interpolated identifiers in the same way.

NtIdent(Ident, IdentIsRaw)

This identifier (and its span) is the identifier passed to the declarative macro. The span in the surrounding Token is the span of the ident metavariable in the macro’s RHS.

Lifetime(Symbol, IdentIsRaw)

Lifetime identifier token. Do not forget about NtLifetime when you want to match on lifetime identifiers. It’s recommended to use Token::{ident,uninterpolate} and Parser::token_uninterpolated_span to treat regular and interpolated identifiers in the same way.

NtLifetime(Ident, IdentIsRaw)

This identifier (and its span) is the lifetime passed to the declarative macro. The span in the surrounding Token is the span of the lifetime metavariable in the macro’s RHS.

DocComment(CommentKind, AttrStyle, Symbol)

A doc comment token. Symbol is the doc comment’s data excluding its “quotes” (///, /**, etc) similarly to symbols in string literal tokens.

Eof

End Of File

Implementations

impl TokenKind

pub fn lit(kind: LitKind, symbol: Symbol, suffix: Option<Symbol>) -> TokenKind

pub fn break_two_token_op(&self, n: u32) -> Option<(TokenKind, TokenKind)>

An approximation to proc-macro-style single-character operators used by rustc parser. If the operator token can be broken into two tokens, the first of which has n (1 or 2) chars, then this function performs that operation, otherwise it returns None.

pub fn similar_tokens(&self) -> &[TokenKind]

Returns tokens that are likely to be typed accidentally instead of the current token. Enables better error recovery when the wrong token is found.

pub fn should_end_const_arg(&self) -> bool

pub fn is_delim(&self) -> bool

pub fn open_delim(&self) -> Option<Delimiter>

pub fn close_delim(&self) -> Option<Delimiter>

pub fn is_close_delim_or_eof(&self) -> bool