rustc_ast::util
Module comments

Structs

Comment

Enums

CommentStyle

Functions

beautify_doc_string

Makes a doc string more presentable to users. Used by rustdoc and perhaps other tools, but not by rustc.

may_have_doc_links

A fast conservative estimate on whether the string can contain documentation links. A pair of square brackets [] must exist in the string, but we only search for the opening bracket because brackets always go in pairs in practice.