Module rustc_ast::util::parser

Enums

AssocOp

Associative operator with precedence.

ExprPrecedence

Fixity

Constants

PREC_CLOSURE

PREC_FORCE_PAREN

PREC_JUMP

PREC_PAREN

PREC_POSTFIX

PREC_PREFIX

PREC_RANGE

Functions

contains_exterior_struct_lit

Expressions that syntactically contain an “exterior” struct literal i.e., not surrounded by any parens or other delimiters, e.g., X { y: 1 }, X { y: 1 }.method(), foo == X { y: 1 } and X { y: 1 } == foo all do, but (X { y: 1 }) == foo does not.

needs_par_as_let_scrutinee

Suppose we have let _ = e and the order of e. Is the order such that e in let _ = e needs parentheses when it is on the RHS?

prec_let_scrutinee_needs_par

In let p = e, operators with precedence <= this one requires parentheses in e.